pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.1",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "openai>=1.92.2",
    "pandas>=2.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-telegram-bot>=22.1",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "click>=8.1.7",
    "flask-apscheduler>=1.13.1",
    "flask-login>=0.6.2",
    "flask-wtf>=1.1.1",
    "marshmallow>=3.20.1",
    "flask-limiter>=3.5.0",
    "argon2-cffi>=23.1.0",
]
